server-python/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.4
distro==1.9.0
faiss-cpu==1.8.0
Flask==3.0.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
jsonify==0.5
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.45
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
langsmith==0.1.50
MarkupSafe==2.1.5
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.23.4
orjson==3.10.1
packaging==23.2
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.2
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
simple-websocket==1.0.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
wsproto==1.2.0
yarl==1.9.4
